Add files via upload
PatternMatching looks for the occurrences of a particular string within a length of DNA. Unlike the PatternCount algorithm (in SymbolArray.py) , it records which indexes those strings appear at. This might be used to search for genes, or occurrences of a putative binding site. In the UCSD Bioinformatics course, the algorithm is used to check whether a possible DNAa binding site appears anywhere else other than the origin of replication.